fix(engine): NPE in PendingMessageSubscriptionChecker
A null check is added to visitPending() method. When a subscription not found, a warning logs the situation.
Also, a comment added to clarify the situation further. Writing a test for the race condition is not trivial since the
written test itself can be non-deterministic. E.g. we create a couple of thousand subscription records in the state
and remove one asynchronously in the test can lead to a flaky test as well.

(cherry picked from commit 7198b69252c047cf550c70c9f11906960842f0cf)