fix(raft): prevent members from demoting too early
This changes `allMembers` so that during join consensus where a member
can be in `oldMembers` and `newMembers` but with different types, we use
the "higher" type.
This fixes a bug where demoting the leader to `PASSIVE` would result in
the leader transitioning too early, right when the joint consensus
configuration entry that contains the leader as `ACTIVE` in `oldMembers`
but `PASSIVE` in `newMembers` is committed.

A demoted leader should only step down once it has committed the final
configuration entry where it only appears as `PASSIVE`. With these
changes, `RaftClusterContext` uses the "highest" type of any member.