fix(engine): provide suppliers to ArrayProperty
The ArrayProperty used to take a single object as inner value. This has
been changed to now require a Supplier. All creations of the
ArrayPropery had to be adjusted accordingly.

It required a Supplier so we are able to create a separate iterator
everytime iterator() is called. This is done so we can read the values
of the array in a thread-safe manner.