test: get Identity version during runtime
The Identity version was hardcoded. This means it would get out of sync
soon, unless we manually update the version everytime.

We can retrieve this version during runtime through one of the classes
in the Identity SDK. As a result we'll always run these tests with the
tag that's defined in our pom file.

(cherry picked from commit 95d3eb04ebb0361e1cef1eddb2a631254d674267)