fix: Fix flaky test shouldResumeExporting in ExportEndpointIT
Since the cluster being used has more than one partition, it might
happen that the startProcess command happens before than the remaining
partitions have received the process. That is because when a process
gets deployed, it will be first sent to partition 1 and then forwarded
to the remaining partitions.

Since the test here is only trying to assert that exporting gets
resumed, then it is ok just to publish a message that will generate
a record instead of deploying a process.

(cherry picked from commit 92006104da16326679a594d75e6b88ad7383a874)