fix(atomix): fix DNS/UDP resource leaks
Fixes a UDP resource leak associated with the `DnsAddressResolverGroup`
usage. The group is meant to be reused and shared, as the channels will
not close any associated resolvers/groups. By creating a new group for
every channel, especially with transient channels, we end up with a new
group that is created and garbage collected, but the underlying socket
is never closed (since as far as the channel is concerned, it doesn't
"own" the group).

While we could create one group per channel and ensure it's closed along
with the channel via a promise, the group is thread safe and seems
perfectly reusable, so let's do that and make full use of the underlying
cache.