fix(engine): don't consider banned instances to be active
When we delete a resource we check for active process instance. An
active process instance is an instance that can still progress be
completed or terminated. A banned instance is neither of those things.
It is a broken process instance which is unrecoverable. A banned
instance should not block deletion of a process definition.