fix(engine): reject correlation on message partition
The `MESSAGE_SUBSCRIPTION/REJECT` command should be written on the
message partition so that it receives feedback after trying to correlate
a message.

As far as I can see, the `REJECT` was always written on the process
instance partition which appears to be incorrect.

(cherry picked from commit dd096d326e59f707932e1766ce1467432e7e8630)