fix(raft): reliably transition local member on updates
Instead of trying to figure out if the local member needs to transition
into a new role by looking at previous and new configuration, just
compare the localMember role before and after applying new config.
If the role is now higher than before, the configuration change
promoted the local member and we should start a transition.