refactor(dist): rewrite ClusterEndpoint using RestControllerEndpoint
WebEndpoint do not allow to parse complex json body. For scale request
we have to accept a list of broker ids. This was not possible with
the `@WriteOperation`.

The difference is that this new endpoint is only available over rest, while
`WebEndpoint` is also available over jmx. But we don't have to expose it over
jmx.