feat: resource deletion events provide tenant information
Make The tenant ownership of the deleted resource clear
from the resource deletion event by providing the tenant
id in the ResourceDeletionRecord.

For resource deletion commands, the tenantId property of
the ResourceDeletionRecord is set to an empty string.
This is fine, as we determine the tenant ownership from
the user's list of authorized tenants.