feat(topology): return an error response with pre-defined error code instead of exception
When the operations return an exception, it is swallowed by the transport. Client only
see the error message. This makes it difficult to map it to a proper http status code
in the endpoint.

In this commit, we defined some specific error codes that can happen when processing
a topology change request. Any other will be converted to a INTERNAL_ERROR. We can add
more specific error codes as needed.