feat(engine): optimize deletion resource lookup after command distribution
When a new ResourceDeletionRecord command is processed, the resource is
looked-up using the user's authorized tenants. When the resource if found,
the tenantId is set on the ResourceDeletionRecord, so that the follow-up
event may provide the tenant for which the resource was deleted. This
information can be used to optimize the resource look-up time when the
command is distributed to the follower brokers in the partition by using
the determined tenantId instead of using the authorized tenantIds again.