fix(raft): read latest snapshot from snapshot store to update the cached snapshot
The current snapshot is updated two ways.
1. When a new snapshot is committed by the SnapshotStore, and the listener is executed
   asynchrously.
2. When the follower recieves a snapshot, the current snapshot is updated immediately after
   it is committed.

Since the listener from step 1 is scheduled on the Raft context, it might be executed after
another InstallRequest is processed. If that happens, the listner might overwrite the
currentSnapshot with an older one. Eventually the cached snapshot will be updated to the latest
one. But there can be short time when it is not reflecting the latest one. This can
cause issues in append requests and install requests.

To prevent this race condition, we update the cached snapshot by directly reading it from the
persisted snapshot store. This way the out of order execution of the listener will not
overwrite the current snapshot info. The listener is just a way to notify that a new snapshot
is available in the snapshot store.