feat(protocol): adding new command intents for the timeout update
I've added the following intents:

UPDATE_TIMEOUT
This command will look for the Job in the state and for it's deadline.
If it can't find one of those it will return a NOT_FOUND rejection.
Otherwise it will write the TIMEOUT_UPDATED event which will modify the
deadline in the state.

TIMEOUT_UPDATED
This event will update the JOB_DEADLINES ColumnFamily. Since the key of
this ColumnFamily is deadline-jockey we can't just update the existing
one. Instead we will need to remove the existing entry for this Job, and
 insert a new value. For this we must have access to the old deadline.
 This is stored in state and should available through the JobRecord.