test(gateway): prevent test failures when topology update runs late
This adds a waiting step to ensure that updating the topology takes
effect before making the first requests.
Alternatively, we would have to use the controllable actor scheduler to
wait for the update job to run or change the topology manager api to
return a future instead of void. Both of the alternatives appeared to be
more work with little benefit.