refactor(engine): change PI cancel processor to TypedRecordProcessor
The ProcessInstanceCancelProcessor was using some intermediary
ProcessInstanceCommandHandlers object to process the command. This was
the only command handler that was existed. This was still a legacy
mechanism. In the past it also handled the creation of workflows and
updating of a payload.

Since all other processors register themselves as a TypedRecordProcessor
 I decided to align the ProcessInstanceCancelProcessor with this
 behaviour as well. As a result, we can get rid of the entire
 ProcessInstanceCommandHandlers concept.

(cherry picked from commit 8eb4ce09b8e6d4fb048692cf1cdae34c59193b25)