test(qa): increase timeout to prevent flakiness in ScaleDownBrokerTest
Eventhough we reduced the swim timeouts, the cluster size in gateway topology
is updated when it receives cluster topology gossiped by ClusterTopologyGossiper.
Since it uses a hard-coded value of 10s, sometimes 10s is not enough to receive the
update. We increase the waiting time to give enough time for the gossiper to propagate
new cluster topology. A better way is to make the delays in ClusterTopologyGossiper
configurable. This will be done later.