feat(protocol): add process instance migration record
This adds the process instance migration record, record value, intent,
and type mapping.

To keep the commit small, I've only added an empty record value. The
properties of the record value can be added in a separate commit. The
only property it has is processInstanceKey because it is process
instance related. This is equivalent to PROCESS_INSTANCE_MODIFICATION.