feat(protocol-impl): add ProcessInstanceMigrationRecord
Implements ProcessInstanceMigrationRecordValue which currently only
requires the processInstanceKey property. Typically this property will
contain the same value as the record's key, but ProcessInstanceRelated
records also provide this explicit property for easy of use and
understanding.

In the gateway, we'll need to make sure to fill this property with the
same value as the record key based on the request data.

Note that the processInstanceKey is required, and so it is included in
both the regular and the "empty" record.