feat(topology): enable coordinator to cancel an ongoing change
This is an inherently dangerous operation as there is no coordination
among the nodes. To increase the safety, coordinator tries to check with
the last member that is trying to apply the operation before cancelling.
The idea is that cancel should be only called if the operation is stuck
and not able to make progress on its own.