feat(protocol): add PI migration record properties
Adds the properties target process definition key and mapping
instructions to the process instance migration record.

I've chosen to add these properties to the record itself, instead of to
a MigrationPlan property. This additional layer is useful in the API,
but not in the record structure. In the record structure it's better to
keep the properties flattened.

I've also allowed to set the source and target element ids of a mapping
instruction to "" empty string because this may become useful in the
future. In the initial iteration it will not serve any purpose though.