refactor(engine): rename bpmn processors
The name of the ProcessEventProcessors class no longer fits to what it
does (probably already since around v1). Nowadays, it provides the
BpmnStreamProcessor which is an additional layer of processors splitting
the processing of the ProcessInstanceRecord into many different
processors, each focused on its own Bpmn element.

By naming this class BpmnProcessors and changing the name to
addBpmnStreamProcessor this becomes somewhat more evident. Of course it
still misses the fact that this is just splitting up the
ProcessInstanceRecord processing, but that is hard to cover without
changing the names of many classes.

BpmnProcessors is a name that fits the usage well.