refactor(gateway): respect client stream's transport state
The gateway now respects the client stream's transport state. That is,
if the stream is not ready to receive messages, it will return an error
back to the transport. The transport can then retry with another
logically equivalent worker, or yield the job back to the broker.

A new error type is introduced to represent this case. This can later be
used to report a proper protocol error identifying the issue.

Additionally, error handling is improved in the gateway on push. We
separate handling errors from serialization with pushing, and also
properly report it instead of simply throwing an exception (which is
logged as an unexpected failure by the actor).