test(engine): should restore linked form
This test fails because of a bug.

It deploys a form alongside a process and creates an instance with a
user task that links the form. After restart, the state must be restored
in such a way that the form can still be linked, but this is not the
case.