fix(engine): support form records
This fixes a bug where form records were not replayed/processed. It was
not noticed because form records don't have commands yet and so
processing was not needed yet. Event applying goes through the event
appliers, so replay was missed.

It solves the bug by including the Form value type in the
SUPPORTED_VALUETYPES enumset.