test(engine): write migrated event
When a migration is successful, we must write the MIGRATED event with
the same data as the command.

Note that we cannot test response writing in the engine tests. This is
because the command is not written as a request that can be responded to
(the requestId and streamRequestId are missing).

It would be great to support this at some point, but I'd rather make
progress on the main topic now. We'll be able to test the response in
the IT where the real java client makes a request that moves through the
gateway to the broker and back.