test(engine): verify element migrated applier
The ProcessInstanceElementMigratedApplier has to do one specific thing,
so it makes sense to write a test case for it. It should update some
specific properties of the element instance.

For a process instance only the flow scope key cannot be adjusted (it's
flow scope will never change during a migration because no sub process
can contain a process instance). Therefore, I've added an additional
test case that shows that this should happen using a service task.