fix(topology): use a fixed timestamp when initializing topology
This is to ensure that when multiple brokers initialize using static initializer,
the resulting topology is equal. Otherwise even if the topology is same, equals check fails
because of different timestamp.

Instant.MIN is less than OffsetDateTime.MIN. So converting Instant.MIN in the ClusterEndpoint
fails with Invalid value. To address this we explicitly check if the value is Instant.MIN before
mapping to OffsetDateTime.