fix(raft): do not include local member in quorum if it is no part of the configuration
This can happen if the local member is the leader and it is leaving the replication group.
Without this fix, it marks the configuration as committed even if it is not committed to
the quorum in the new configuration. For example, 0,1 and 2 are the current members. 0 is the
leader and trying to leave. It should only commit the new configuration if it has been committed to
both 1 and 2. But due to the bug, 0 includes itself in the quorum and commits even if it was only
written to 1.