feat(engine): migrate the process instance
This migrates the process instance itself by appending the
ELEMENT_MIGRATED event with adjusted data. The associated event applier
updates the data in the database.

I've already added some if cases where we're unable to find the process
instance or unable to find the process definition, but the handling is
not fully correct yet. We can do that in the upcoming Rejection issues.