feat(client-java): add command steps for Process Instance Migration
Steps are structured in a way that they force users to follow required order.
E.g. targetProcessDefinitionKey needs to be provided before mappingInstructions.
Also, addMappingInstruction method is duplicated in both step 2 and final step.
This is because we still want to be able to add intermediate steps to the command
chain using step 2 interface.