feat: Azure Backup Store basic functionalities
This commit implements the basic Azure Backup store such as configuration
being able to configure the Store, being able to save contents and manifests,
get status of backups, list backups, delete backups, and restore backups.

Currently the only tests used are the ones provided by the BackupStoreTestKit,
and a specif test for the Azure Backup configuration.

The other classes are missing specialized tests for Azure, and these will be
added in following PRs.