fix(auth): don't add issuedAt jwt token claim
Don't add a JWT issuedAt claim when creating an auth
token. Zeebe is currently not signing its JWT tokens
so an issuedAt claim is not necessary. Furthermore,
since we verify the JWT token for the presence of
certain claims, an  claim might cause an
exception to be thrown if the JWT verification is not
configured correctly (note: issuedAt verification
is done automatically).