test(auth): add a test case for jwt issue date verifcation
Add a test case to ensure that JWT verification
doesn't fail if the issuedAt date is set in the
future. We want to ensure this is possible in
case the clocks between different Zeebe components
(ex. Gateway and Broker) are slightly out of sync.

In that case, some leeway should be allowed, or
the issuedAt verification disabled.