fix(auth): disable jwt issued at verification
Disable the verification of the JWT issuedAt claim.
This claim is verified automatically if not disabled.
By default, the leeway for successful verification
is 0, so even a slight differnce between the
clocks of different components may cause the
verification to fail.

In the Zeebe case, we don't sign the JWT token
so verifying the issuedAt claim is not relevant.

(cherry picked from commit 07d3c0178e2dd563961e2c84e904c81154f87041)