refactor(qa): relax jobs consumed condition
Relaxes the condition to await all expected jobs. Instead of looking for
a specific sequence, we just assert that all expected jobs were
activated, in any order. This is sufficient for this test, since it's
difficult to get a proper ordering with streaming and polling in
parallel.

Receiving the jobs in sequence is also not part of the worker contract,
so no real need to test it here.