fix: cleanup mem leak
ByteBuf allocated by Channel, uses internally a direct memory pool and reference counting.
The reference counting and error log allowed us to detect that there was a memory leak, since
we didn't closed the ByteBuf before.

This commit uses futures to clean up the buffer when the requests were send or in case of error that
it is still released.