test(engine): migrate from v1 to v2
The test was migrating from v2 to v2, because the instance was created
without instructing an explicit version. Since both versions were
already deployed, the instance was created of the latest version.

This fixes the test, to actually migrate from v1 to v2.

The test did not fail because it does not assert the assumption that the
instance was created on v1 of the process definition. I see no need to
add this assertion still as it does not do more than verify that what is
already defined above.