refactor: changes for Broker context to also share the commandApiMessagingService
The nature of the flaky test seemed to be that the disconnect method was
missing closing the commandApiMessagingService in the BrokerStartupContextImpl
as well. Which caused that we could still reach the broker even after
disconnecting. The test would pass most of the time because the timeout limit
was smaller than the round trip time of the request, and we would get a 504
timeout.