feat(engine): migrate all element instances
This migrates all element instances of a process instance. It traverses
through the tree of element instances in a breadth-first fashion,
starting with the direct child instances of the process instances.

It could have been fine to only migrate the direct child instances, and
ignore the other descendants in this first iteration. However,
eventually we must support this anyways, so I've already added this
capability to the code. Support for migrating nested structures in
process instances will be controlled through validation to safeguard
users from existing bugs. This makes it easy to add support for nested
structures later because we can focus on test cases instead of
implementation details.

Traversal is done using a queue to avoid a potential stackoverflow.

Determining the target element id highlighted an interesting aspect to
the current design of the mapping instructions: users can define mapping
instructions with the same source element id. This would mean it's
unclear to the engine which of the matching mapping instructions to use.
So we need some validation that verifies that the source element id of
each mapping instruction is unique.