refactor(engine): reuse element instance migration
The process instance is just another element instance, we don't have to
migrate it separately from the other element instances. Instead, we can
re-use the element instance migration logic for the process instance.

Eventually this will make it easier to migrate associated things like
variables for the process instances and the descending element
instances.

In order to do this refactoring, we have to determine the target element
id of the process instance slightly differently because users don't
provide a mapping instruction for the bpmn process id (the element id of
the process instance element). This is an acceptable trade-off, because
we win the reuse for migration of event subscription in the scope,
variables in the scope, etc.