feat(engine): support user task processing
Decide on native user task or user task job in a new processor for BPMN User Tasks.
If job properties are present, delegate to the existing processing behavior.
Otherwise, establish a new native user task processing that can create native user task records and state.
The user task properties are tracked in a new ExecutableUserTask object that allows the differentiation.