build: introduce quickly mvn flag
Introduces a single flag to skip all non-essential goals. Including
skipping the assembly. This is useful for local compilation where a
polished, compressed and archived artifact is not necessary.

In the future, we can hopefully also skip the flatten plugin, and then
this could also be part of the quickly flag.