feat(gateway-protocol-impl): add MigrateProcessInstance rpc to gateway service config
The MigrateProcessInstance rpc added to the second config where requests
are not retriable on DEADLINE_EXCEEDED. This is because even the server
is not able to process the request before the deadline, it might still
process the migration. In that case, we need to prevent processing
the migration command twice.