test: reject migration of unsupported element type
This adds test cases for most element types to show that a migration is
rejected when that element is active.

Not all element types are covered, because it's not possible to make the
engine wait at certain elements. For example, the engine activates and
completes a START_EVENT almost instantly. While it may be possible, it's
exceptionally difficult to test for. Instead, most other cases are
covered. Additionally, the implementation is very straight-forward, as
you'll see in the upcoming commits.