feat: reject migration of unsuported element type
Adds the ability to reject migration command when the process instance
has an active element of a type that is not supported for migration.

Failure handling is implemented by throwing an exception while migrating
each of the element instances of the process instance. This exception
can be handled as an expected error in the tryHandleError method. This
ensures the processing is rolled back (no state changes and no records
appended) for safe error handling. In the tryHandleError we can then
respond with a rejection.

One downside to this method is throwing and catching an exception, which
can be costly. An alternative is available by validating the active
elements before migrating the element instances but this costs an
additional iteration through the tree of element instances or requires
to keep all encountered element instances in memory. Both are worse than
throwing an exception and handling it as expected.

The assertion is made by checking whether the element is supported or
not, which is defined statically using a simple EnumSet like in
ProcessInstanceModificationModifyProcessor.