refactor: construct error message in exception
The constructor of the exception is a much nicer place to construct the
error message. This has several benefits.

At runtime, you'll be able to see all information in the
message about the exception rather than only when the exception is being
handled by the processor.

It reduces eye-strain when reading through the error handling of the
processor.

It helps document inside of the exception what its purpose is.