fix: do not delete jobs from backoff cf
Do not delete jobs from backoff CF if they are still existing, or have a correct recurring time set.

Previously we checked the wrong property which was always not the same as in the state, this caused us
to delete jobs from the backoff column family. This means jobs were stuck in the failed state forever.