refactor: rework existing JobBackoffMigration tests
Previous test used directly the column families to set up and update value,
which is really error prone since underlying or usage of column families
can change and this is not reflected in these test. Meaning they will not fail, as we saw.

This commits couples the test with the actual JobState in order to set up the column family
data correctly, and update it in an expected way.

In this case the migration can also work properly and we can verify whether clean up
works ok or not