feat(engine): add user task creation event applier
Writing native user task events to the log is reflected in the state as well.
Upon the CREATING event, the related user task record is inserted to the state.
Upon the CREATED event, the record in the state is updated.
This ensures we can retrieve the user task data from the state when required later,
e.g. for cancelation or completion.