deps: add identity-spring-boot-autoconfigure dependency
Add the identity-spring-boot-autoconfigure dependency to Zeebe.
The dependency will provide an IdentityConfiguration Spring bean
if the Identity class is on the classpath. This will decouple the
identity-sdk configuration properties from Zeebe, allowing the
projects to be developed more independently.

This commit also adds the identity-sdk to the Zeebe dist and
broker modules since the IdentityConfiguration class is now
passed through the classes of these modules.

This commit also marks the identity-spring-boot-autoconfigration
dependency as used in the Maven Dependency plugin since the
Spring Boot autoconfiguration classes are not explicitly used.

The com.squareup.okio:okio dependency is excluded from the
dependency convergence checks. This issue needs to be resolved
by Identity (and a GitHub issue already exists for that).