feat(gateway): instantiate an identity interceptor using identity configuration
Allow for an instance of the  IdentityInterceptor class to be
instantiated using an instance of the IdentityConfiguration
class. This will allow the Gateway to accept Spring beans of
the IdentityConfiguration class that will be provided by the
identity-spring-boot-starter dependency.

If users already configured Identity through the Zeebe config
properties, then this configuration will take precedence over
the dedicated Camunda Identity configuration properties.

A backwards-compatible Gateway constructor is added that
passes null for the IdentityConfiguration instance. The
Identity component is optional, and by default it is disabled,
so the old constructor can still be (validly) used in tests.