feat(broker): pass identity configuration in broker classes
Allow the Broker to pass along an instance of the IdentityConfiguration
class. The IdentityConfiguration instance will be used to create an
instance of the IdentityInterceptor when an EmbeddedGateway is used in
the Broker. The goal of this change is to allow an IdentityConfiguration
Spring bean to be passed from the StandaloneBroker Spring Boot applicaction.