feat(state): allow retrieval of variables at scope
Variables are stored in the state by:
  [scopeKey, name] -> [key, value]

So in order to retrieve variables we have to do so by scope key.
Therefore, it's not possible to retrieve all variables for a process
instance at once.

Instead, we need to have a way to retrieve variables at a specific
scope. However, most methods of the VariableState either:
- get a single variable value
- or get a merged variables document (like json a document)

No methods were available to retrieve the details of a variable, let
alone of all variables at a specific scope. So, I added it now.

A VariableRecord requires additional data that is not (currently) stored
in the VariableState like processDefinitionKey and processInstanceKey.
So, it's not possible to create them from just the details in the state
and return these. Instead, I've added a record `Variable` that allows us
to easily wrap the details of a variable.

Note that the visitor is not safely visiting the variables, so we have
to copy the name and value buffers on each visit.