feat(engine): append variable migrated events
For each of the variables at the scope of an active element instance, we
need to migrate those variables to the new process definition.

We can use the new getVariablesLocal method from the variableState to
retrieve all variables at the scope of the element instance.

The variableState does not store all details needed to create a
VariableRecord, but we can provide that information from the migration
in progress and the active element instance.

In the tests, I've validated the migration of multiple variables to
ensure the buffers are copied and we don't end up with reused buffer
data (e.g. 2 Variable MIGRATED events but both describing the same
variable).