feat(engine): unset value in variable migrated
The value in a variable is user-defined and can be large. It's easy to
reach the MAX_MESSAGE_SIZE (or max record batch size, default 4MB) if we
were to include the variable's value in the Variable MIGRATED event.

As it is not strictly necessary, we can exclude it. However, that has
some complications.

Unsetting the value means explicitly setting it to `null`. We have no
other way to indicate that the value is not set. This means that the
standard event applier in use for variable records can no longer be
used. It attempts to update the persisted variable's data with exactly
that what is contained by the record. For MIGRATED variables this means
overwriting the value to null in the state, which we should not do.

In the same way, users should ensure not to copy all information from
the Variable MIRGRATED event. They should be aware that the value is
unset to null.