fix(topology): propagate unhandled initialization errors
Instead of silently accepting any exceptions returned from an
initializer and continuing in the chain, we want to propagate the
exception until it is handled by `recovery` or the initialization chain
fails entirely.

All initializers so far always complete successfully and not
exceptionally, except for the `FileInitializer` when the file is
corrupted. In that case, we want the exception to propagate so that the
coordinator which doesn't use `recover` fails to initialize.