feat(engine): add user task lifecycle state
Store a native user task's internal lifecycle state in Zeebe's state.
The state is not a property of the native user task record and is not exported.
The lifecycle state is an implementation detail that is not used externally.

Upon creating a native user task, it has the state CREATING.
The state can be updated with the related `updateUserTaskState` method.