test: refactor tests to use identity configuration directly
Ensure that integration tests use the dedicated Identity
configuration properties instead of the Zeebe properties
for Identity.

Pass a null for the IdentityConfiguration bean in the
Gateway instances used in tests where Identity is not
used.

Improve the Identity-related IT tests to obtain the
Identity version being used in a more robust way. The
new method I add supports snapshot versions as well.