refactor(msgpack): require a hint for the number of declared properties
The size hint is used for the array holding declared properties.
This is especially useful for large records with many declared
properties which would previously need to resize the array.
It also helps small records with less than 10 declared properties, the
default for `ArrayList` when not providing a size hint.

While it would be preferable to not have to provide the number and
instead provide the list of declared properties directly, initialization
order forbids us to use the instance fields before calling the super
constructor.