feat(engine): Reject process instance migration if mappings refer to non-existent element id
Please note that I iterated over `mappingInstructions` list directly.
It is because we add process instances to the mapping inside `mapElementIds`
method but users do not provide mapping for process definitions in mapping
instructions.

Also, I preferred to use a common exception for both cases but allow to
specify if the mapping is wrong for source or target process definition.