feat(engine): Reject migration if event is triggered
Adjust the process instance migration processor to reject a migration command if an event trigger is created (i.e. with
intent TRIGGERING). The event trigger indicates a concurrent command for the process instance. It is created when a job
is completed or a BPMN event is triggered (e.g. message, timer, or signal).