refactor(engine): reduce user task state access
Upon checking a user task command, the user task record is fetched from the state to validate authorizations.
The same record is fetched from the state afterward in the command again to manipulate it.
To reduce the number of fetch invocations, the checker returns the fetched record in success cases.