feat(engine): Reject process instance migration if element is nested in a sub process
Assume we try to migrate Process 'A' > Service Task 'B' into target definition
Process 'C' > Service Task 'D'. We first update values of Process 'A' inside
`tryMigrateElementInstance` method. That means, the elementId of Process 'A'
becomes 'C'. Afterwards, when we try to migrate Service Task 'B',
the retrieved sourceFlowScopeElement's elementId of B becomes 'C'.
Therefore, we compare that `expectedFlowScopeId` with the one in target process definition
which is `actualFlowScopeId`.