feat(engine): Reject migration if active sequence flow
Adjust the process instance migration processor to reject a migration command if the process instance has an active
sequence flow. An active sequence flow indicates a concurrent ACTIVATE command for the process instance. The number is
increased when a sequence flow is taken and an ACTIVATE command for the next element is written. The number is
decreased when the ACTIVATE command is processed.