fix(topology): ensure topology messages are processed in order
When gossiper directly updates its local GossipState, the order of execution can differ
because the ActorFuture:onComplete enqueues the task at the end of the queue. As a result
during initialization, sometimes a valid topology gets overwritten by an uninitialized one.
To simplify handling this, we always updated the gossiper from the ClusterTopologyManager
directly. This ensures correct ordering, whether Gossiper is running on same exeuctor or not.