test(expression-language): Correct test for special characters
The test case should verify that a special character in an expression with a string literal is not escaped.

But the expression was defined as a static value and not a FEEL expression.

Modify the expression to align with the intent of the test case.

(cherry picked from commit 6eb89da0f7021122cc372906d01faea43d6211f2)