fix(topology): use separate actors for gossip and manager
While we were previously using the same actor for gossip and topology
manager, the order of applying topology updates could be confused
because the gossiper would reorder updates because actor run prepends
jobs.
This behavior was wrong and difficult to keep track of. To fix this bug
and simplify the interactions in the future, we now use two separate
actors for both. This ensures that updates to the gossip state are
ordered by the usual actor scheduler semantics.

Co-authored-by: Deepthi Devaki Akkoorath <deepthidevaki@gmail.com>
(cherry picked from commit d2bcf8d666833284ee5246e692352d3dfb06f92e)